Consolidate producers under single dir
