+++ src/producers/beamhandler.ts
@@ -1,6 +1,6 @@
-import * as msg from './msg';
+import * as msg from '../msg';
+import { VexmlMessageReceiver } from '../types';
 import { NodeHandler, NodeHandlerCtx } from './nodehandler';
-import { VexmlMessageReceiver } from './types';
 
 export class BeamHandler extends NodeHandler<'beam', NodeHandlerCtx<'beam'>> {
   sendMessages(receiver: VexmlMessageReceiver, ctx: NodeHandlerCtx<'beam'>): void {
