+++ src/producers/measurehandler.test.ts
@@ -1,9 +1,9 @@
-import { DEFAULT_CONFIG } from './di';
+import { DEFAULT_CONFIG } from '../di';
+import { NoopReceiver } from '../noopreceiver';
+import { MeasureEndMessage, MeasureStartMessage } from '../types';
+import * as xml from '../xml';
 import { MeasureHandler } from './measurehandler';
 import { NoopHandler } from './noophandler';
-import { NoopReceiver } from './noopreceiver';
-import { MeasureEndMessage, MeasureStartMessage } from './types';
-import * as xml from './xml';
 
 describe('MeasureHandler', () => {
   let receiver: NoopReceiver;
