+++ src/producers/measurehandler.ts
@@ -1,8 +1,8 @@
-import * as msg from './msg';
-import { MultiReceiver } from './multireceiver';
-import { NamedNode } from './namednode';
+import * as msg from '../msg';
+import { MultiReceiver } from '../multireceiver';
+import { NamedNode } from '../namednode';
+import { VexmlConfig, VexmlMessageReceiver } from '../types';
 import { NodeHandler, NodeHandlerCtx } from './nodehandler';
-import { VexmlConfig, VexmlMessageReceiver } from './types';
 
 export class MeasureHandlerError extends Error {}
