+++ src/producers/nodehandler.ts
@@ -1,5 +1,5 @@
-import { NamedNode } from './namednode';
-import { VexmlMessageProducer, VexmlMessageReceiver } from './types';
+import { NamedNode } from '../namednode';
+import { VexmlMessageProducer, VexmlMessageReceiver } from '../types';
 
 export type NodeHandlerCtx<T extends string> = {
   node: NamedNode<T>;
