+++ src/producers/notehandler.ts
@@ -1,7 +1,7 @@
-import * as msg from './msg';
-import { NamedNode } from './namednode';
+import * as msg from '../msg';
+import { NamedNode } from '../namednode';
+import { NoteMessageHead, VexmlConfig, VexmlMessageReceiver } from '../types';
 import { NodeHandler, NodeHandlerCtx } from './nodehandler';
-import { NoteMessageHead, VexmlConfig, VexmlMessageReceiver } from './types';
 
 export class NoteHandler extends NodeHandler<'note', NodeHandlerCtx<'note'>> {
   private config: VexmlConfig;
