+++ src/producers/todohandler.ts
@@ -1,5 +1,5 @@
+import { VexmlMessageReceiver } from '../types';
 import { NodeHandler, NodeHandlerCtx } from './nodehandler';
-import { VexmlMessageReceiver } from './types';
 
 export class TodoHandler extends NodeHandler<string> {
   sendMessages(_receiver: VexmlMessageReceiver, ctx: NodeHandlerCtx<string>): void {
