+++ src/renderer.ts
@@ -3,7 +3,7 @@ import { BoundingBox, Clef, StaveModifierPosition } from 'vexflow';
 import { Attributes } from './attributes';
 import { CodePrinter } from './codeprinter';
 import { Notations } from './notations';
-import { Producer } from './producer';
+import { Producer } from './producers/producer';
 import { CodeTracker, NoteMessage, VexmlMessage, VexmlMessageReceiver } from './types';
 
 export type RendererOptions = {
